Update namings and default behaviors
