+++ Sources/GlyphixTextFx/GlyphixTextModifiers/GlyphixText+Blur.swift
@@ -1,5 +1,5 @@
 //
-//  GlyphixText.Blur.swift
+//  GlyphixText+Blur.swift
 //  GlyphixTextFx
 //
 //  Created by LiYanan2004 on 2025/3/10.
@@ -9,7 +9,7 @@ import SwiftUI
 
 extension View {
     /// Sets whether label should disable blur effect.
-    public func glyphixLabelBlurEffectDisabled(_ disabled: Bool = true) -> some View {
+    public func glyphixTextBlurEffectDisabled(_ disabled: Bool = true) -> some View {
         transformEnvironment(\.blursDuringTransition) { blurEffectEnabled in
             blurEffectEnabled = !disabled
         }
